pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "microweldr"
version = "6.1.10"
description = "Convert SVG and DXF files to Prusa Core One G-code for plastic welding"
authors = [
    {name = "Reto Stamm", email = "reto@retostamm.com"},
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["3d-printing", "welding", "gcode", "svg", "dxf", "prusa", "microfluidics", "plastic-welding", "bambu", "core-one"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
    "Typing :: Typed",
]
dependencies = [
    "Pillow>=10.0.0",
    "ezdxf>=1.4.0",
    "lxml>=5.0.0",
    "toml>=0.10.2",
    "requests>=2.25.0",
    "pyserial>=3.5",
    "bambuuzle>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/retospect/microweldr"
Repository = "https://github.com/retospect/microweldr"
Documentation = "https://github.com/retospect/microweldr"
Issues = "https://github.com/retospect/microweldr/issues"
Changelog = "https://github.com/retospect/microweldr/blob/main/CHANGELOG.md"

[project.scripts]
microweldr = "microweldr.__main__:main"
microweldr-usb = "scripts.prusa_usb_terminal:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.11",
    "mypy>=1.15",
    "pre-commit>=4",
    "requests-mock>=1.11",
]

[tool.uv.sources]
bambuuzle = { path = "../bambuuzle", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["microweldr", "scripts"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B017", "RUF043", "SIM115"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "ezdxf.*",
    "lxml.*",
    "serial.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "hardware: marks tests that require hardware connection",
    "integration: marks tests as integration tests",
]
